Docker backup script use docker command.
Using docker-compose directly sometimes errored.

From a blog:

In theory that should work, but it won’t (at least not with Docker Compose 1.20). You’ll end up with an error of ValueError: file descriptor cannot be a negative integer (-1).
You could simplify that by replacing $(docker-compose ps -q postgres) with the name of the container if you know it beforehand, otherwise it will dynamically grab it from whatever you have defined as the service name in your docker-compose.yml.