Fix PG foreign key error on deleting devices
This fixes #214 - deleting devices failed since we moved the postprocessing data to a seperate table, as the foreign key constraint prevented the postprocessing table entry being orphaned.

The solution is to make this relationship `depenedent: destroy` - when a device is removed, we remove its postprocessing info too.

I've also added a test to guard against regressions here and make the behaviour we expect explicit.