Fix CSV sort order bug (#215)
Metrics provided by an HTTP post after postprocessing were showing up out-of-order
in the CSV, apended after the sensor data received directly. This was due to the fact that
the CSV table is built column-wise rather than row-wise, and ruby hashes preserve the order of
insertion, not of their keys.

To fix this, we've added an explicit sort by timestamp at the end, this returns an array rather
than a hash, so have renamed the method and refactored to suit.

Still TODO: better test coverage and some refactoring of this class, as this logic could be made
much clearer.