Skip throttling for admins and researchers
This proved to be a little tricky, as the throttling happens in Rack middleware,
*before* we've authenticated the user, and attempting to move authentication earlier
in the request cycle threw up a load of very weird problems, as well as being less
than ideal - as even throttled requests would have to hit the database.

The solution here is, on authenticated requests, to set a flag in the cache for
the requesting IP (with a timeout) in order to suspend throttling for that IP.
therefore, on the next request, throttling will be disabled.

One drawback of this is that unauthorized users on the same IP as an admin or researcher
will not be throttled - however, this is hopefully a corner case, and I can't think of
any other solution that doesn't involve authenticating users before throttling
(with the problems stated above), or making requests stateful using a session cookie
(which both immediately makes our API less RESTful, and makes the throttling trivial
to bypass by simply not sending the cookie.

Let me know if this is an acceptable solution!