Ensure devices are deleted 24 hours after archiving
This was a bug that was hidden by the existing FK error - when deleting archived devices, we check the date they were created, not archived. This has the result that devices *created* more than 24 hours ago don't have the 24 hour grace period and are deleted immediately.

To fix this, we now store an archived_at date on the device, which is populated on archive, and check this in the deletion job.