Delete archived devices which have a null archived_at date.
For some reason, we've got some legacy devices which never had their archived_at date set, which are causing the deletion job to crash, meaning that neither they (nor archived devices with subsequent IDs) are deleted. This adds a test for the failure, and fixes so that all archived devices with a null archived_at are deleted. This shouldn't cause a problem in future as only old devices have the null archived_at date.