Deferred retry of hardware_info handling when device not found
Since we deployed the devices refactor, a slightly subtle bug has
emerged: if a device sends the hardware_info packet before the user has
completed registration, the device is not found, the info is not saved,
and therefore the device information in the UI is incomplete. This
(hopefully, I like to test on staging with a real device) fixes this by
relying on Sidekiq's built in error handling functionality to retry the
hardware_info message (with backoff) in the event that no device
corresponds to the given key. Sidekiq defaults to 25 retries with
backoff (corresponding to about 3 weeks of elapsed time between the
first and the last try), which should be sufficient:
https://docs.gitlab.com/ee/development/sidekiq/.