Lock device at start of mqtt processing
This will hopefully fix the duplicate component bug without adding the
overhead of an extra query on the device table.

This also meant we need to change our database cleaning strategy in the
tests as we can't nest transactions with different isolation levels.
I've included the database_cleaner gem and used it with the truncation
strategy to fix this.