SPIKE - generalising device identifiers. DO NOT MERGE!
This was an attempt to generalise the way we identify devices (by mac
address) in order to easily incorporate other identification schemes
(Such as meshtastic IDs). It's probably an abstraction too far for our
current purposes (allowing meshtastic devices specifically to ingest
data to the platform), but am keeping here for discussion  on potential
future improvements.